# openapi.yaml
openapi: 3.0.3
info:
    title: MechHub Miniback API
    version: 1.0.0
    description: Standalone backend for MechHub. Authentication is session-based and uses the `mechhub.sid` cookie.
servers:
    - url: http://localhost:3001
      description: Local development server
...
components:
    securitySchemes:
        cookieAuth:
            type: apiKey
            in: cookie
            name: mechhub.sid
    schemas:
...
        ErrorResponse:
            type: object
            required:
                - message
            properties:
                message:
                    type: string
                    example: Not authenticated
        Class:
            type: object
            required:
                - id
                - name
                - description
                - ownerUserId
                - inviteCode
                - status
                - createdAt
                - membershipRole
                - isOwner
            properties:
                id:
                    type: integer
                    example: 1
                name:
                    type: string
                    example: Statics 101
                description:
                    type: string
                    example: Introductory statics class for spring term.
                ownerUserId:
                    type: integer
                    example: 1
                inviteCode:
                    type: string
                    example: 8F1A2C4D
                status:
                    type: string
                    enum: [active, archived]
                    example: active
                createdAt:
                    type: string
                    format: date-time
                    example: 2026-03-23T09:10:00.200Z
                membershipRole:
                    type: string
                    enum: [teacher, student]
                    nullable: true
                    example: teacher
                isOwner:
                    type: boolean
                    example: true
...
paths:
...
    /classes:
        get:
            tags:
                - Classes
            summary: List classes the current user has joined
            operationId: listClasses
            security:
                - cookieAuth: []
            responses:
                "200":
                    description: Class list
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/Class"
                "401":
                    description: Not authenticated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"

        post:
            tags:
                - Classes
            summary: Create a class
            operationId: createClass
            security:
                - cookieAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/CreateClassRequest"
            responses:
                "201":
                    description: Class created
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Class"
                "400":
                    description: Invalid class fields
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "401":
                    description: Not authenticated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "403":
                    description: Only teachers can create classes
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /classes/join:
        post:
            tags:
                - Classes
            summary: Join a class by invite code
            operationId: joinClass
            security:
                - cookieAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/JoinClassRequest"
            responses:
                "201":
                    description: Joined class
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Class"
                "400":
                    description: Invalid invite code
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "401":
                    description: Not authenticated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: Class not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "409":
                    description: User has already joined the class
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /classes/{classId}:
        get:
            tags:
                - Classes
            summary: Get a joined class
            operationId: getClass
            security:
                - cookieAuth: []
            parameters:
                - in: path
                  name: classId
                  required: true
                  schema:
                      type: integer
            responses:
                "200":
                    description: Class detail
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Class"
                "400":
                    description: Invalid class id
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "401":
                    description: Not authenticated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: Class not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"

        patch:
            tags:
                - Classes
            summary: Update a class
            operationId: updateClass
            security:
                - cookieAuth: []
            parameters:
                - in: path
                  name: classId
                  required: true
                  schema:
                      type: integer
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UpdateClassRequest"
            responses:
                "200":
                    description: Updated class
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Class"
                "400":
                    description: Invalid class id or class fields
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "401":
                    description: Not authenticated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "403":
                    description: Only the class owner can update the class
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: Class not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /classes/{classId}/invite-code/rotate:
        post:
            tags:
                - Classes
            summary: Rotate a class invite code
            operationId: rotateClassInviteCode
            security:
                - cookieAuth: []
            parameters:
                - in: path
                  name: classId
                  required: true
                  schema:
                      type: integer
            responses:
                "200":
                    description: Updated class with a new invite code
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Class"
                "400":
                    description: Invalid class id
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "401":
                    description: Not authenticated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "403":
                    description: Only the class owner can rotate the invite code
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: Class not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /classes/{classId}/leave:
        post:
            tags:
                - Classes
            summary: Leave a joined class
            operationId: leaveClass
            security:
                - cookieAuth: []
            parameters:
                - in: path
                  name: classId
                  required: true
                  schema:
                      type: integer
            responses:
                "204":
                    description: Left class successfully
                "400":
                    description: Invalid class id or the class owner cannot leave directly
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "401":
                    description: Not authenticated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: Class not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /classes/{classId}/members:
        get:
            tags:
                - Classes
            summary: List members of a joined class
            operationId: listClassMembers
            security:
                - cookieAuth: []
            parameters:
                - in: path
                  name: classId
                  required: true
                  schema:
                      type: integer
            responses:
                "200":
                    description: Class members
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/ClassMember"
                "400":
                    description: Invalid class id
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "401":
                    description: Not authenticated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: Class not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
    /classes/{classId}/members/{memberId}:
        patch:
            tags:
                - Classes
            summary: Update a member's class role
            operationId: updateClassMemberRole
            security:
                - cookieAuth: []
            parameters:
                - in: path
                  name: classId
                  required: true
                  schema:
                      type: integer
                - in: path
                  name: memberId
                  required: true
                  schema:
                      type: integer
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UpdateClassMemberRoleRequest"
            responses:
                "200":
                    description: Updated class member
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ClassMember"
                "400":
                    description: Invalid ids, role, or attempted owner role change
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "401":
                    description: Not authenticated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "403":
                    description: Only the class owner can update member roles
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: Class or member not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"

        delete:
            tags:
                - Classes
            summary: Remove a member from a class
            operationId: removeClassMember
            security:
                - cookieAuth: []
            parameters:
                - in: path
                  name: classId
                  required: true
                  schema:
                      type: integer
                - in: path
                  name: memberId
                  required: true
                  schema:
                      type: integer
            responses:
                "204":
                    description: Member removed
                "400":
                    description: Invalid ids or attempted owner removal
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "401":
                    description: Not authenticated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "403":
                    description: Only the class owner can remove members
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: Class or member not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
...
    /dashboards/teacher/classes:
        get:
            tags:
                - Dashboards
            summary: Get the teacher's class overview dashboard
            operationId: getTeacherClassesDashboard
            security:
                - cookieAuth: []
            responses:
                "200":
                    description: Teacher class dashboard
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/TeacherClassesDashboard"
                "401":
                    description: Not authenticated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "403":
                    description: Only teachers can access this dashboard
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
...
    /dashboards/classes/{classId}:
        get:
            tags:
                - Dashboards
            summary: Get a teacher-facing class dashboard
            operationId: getClassDashboard
            security:
                - cookieAuth: []
            parameters:
                - in: path
                  name: classId
                  required: true
                  schema:
                      type: integer
            responses:
                "200":
                    description: Class dashboard
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ClassDashboard"
                "400":
                    description: Invalid class id
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "401":
                    description: Not authenticated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "403":
                    description: Only class teachers can access this dashboard
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "404":
                    description: Class not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
...